spectrogram: configurable row rate
Move analyzer sampling to an interval instead of animation frames.

Aggregate multiple samples until the row rate has ticked over.
Aggregation is done with "max" since "average of samples" seemed to
degrade the signal too much.

Updating the canvas is still done via an animation frame. Not so much
because it needs the better timing, but mostly to put it on another
runtime tick.