bins: don't return computed values
These functions are usually called in loops and several of the loops are
run several times a second. Computed values are registered in the active
"effect scope" (usually a component). The many hundred computed values
created for each analyzer report registers them in the parent
component's effect scope. This effectively creates a memory leak. It
might be possible to use `effectScope` to isolate these computed values,
but it looks like it would be awkward.

The new binFor and forBin result functions internally use reactive
values, so they can be wrapped in computed by the caller if they want a
reactive values.

In general, useBins, useFFTBins, and useFFTPixelBins should probably
only be called "once" for each set of (reactive) parameters, or (for
temporary "binners") effectScope should be used to isolate the internal
computed values.