Web Audio composables take reactive audio context
This lets the DemoAudio component react to audio context changes without
calling useNoisyPeriodicBeep inside a watch (creating an effect inside
an effect, thus leaking).

The app doesn't actually bother to update the audio context, but this
makes them a more fully reactive.

This commit also removes useNoisyPeriodicBeep's gain input+output. The
app doesn't need it anymore since it has its own monitor gain. Maybe
this takes away some bit of "completeness", but it should be easy to add
back in if it is ever needed.