show third-party license information
The place-holder data file (`public/licenses.json`) will be over-written
by the license-collection plugin for builds.

`rollup-license-plugin` 2.0.0 uses `.default` to access the function
exported by `get-npm-tarball-url`. This works okay with
`get-npm-tarball-url` 2.0.3 where it was a CJS module that provided a
"default" property in its exports. But it breaks when using
`get-npm-tarball-url` 2.1's ESM that has a default export. So, we
override `get-npm-tarball-url` to be the version `rollup-license-plugin`
was written against.